+++ src/full/Agda/TypeChecking/Serialise/Instances/Highlighting.hs
@@ -93,6 +93,7 @@ instance EmbPrj HP.OtherAspect where
   icod_ HP.MissingDefinition    = pure 13
   icod_ HP.ShadowingInTelescope = pure 14
   icod_ HP.CosmeticProblem      = pure 15
+  icod_ HP.InstanceProblem      = pure 16
 
   value = \case
     0  -> pure HP.Error
@@ -111,6 +112,7 @@ instance EmbPrj HP.OtherAspect where
     13 -> pure HP.MissingDefinition
     14 -> pure HP.ShadowingInTelescope
     15 -> pure HP.CosmeticProblem
+    16 -> pure HP.InstanceProblem
     _  -> malformed
 
 instance EmbPrj HP.Aspects where